+++ litecli/main.py
@@ -11,7 +11,7 @@ import traceback
 from collections import namedtuple
 from datetime import datetime
 from io import open
-from sqlite3 import OperationalError, sqlite_version
+from sqlean import OperationalError, sqlite_version
 from time import time
 
 import click
@@ -260,7 +260,7 @@ class LiteCli(object):
             )
             return
 
-        formatter = logging.Formatter("%(asctime)s (%(process)d/%(threadName)s) " "%(name)s %(levelname)s - %(message)s")
+        formatter = logging.Formatter("%(asctime)s (%(process)d/%(threadName)s) %(name)s %(levelname)s - %(message)s")
 
         handler.setFormatter(formatter)
 
@@ -361,7 +361,7 @@ class LiteCli(object):
         else:
             history = None
             self.echo(
-                'Error: Unable to open the history file "{}". ' "Your query history will not be saved.".format(history_file),
+                'Error: Unable to open the history file "{}". Your query history will not be saved.'.format(history_file),
                 err=True,
                 fg="red",
             )
@@ -452,7 +452,7 @@ class LiteCli(object):
                         if context:
                             click.echo("LLM Reponse:")
                             click.echo(context)
-                            click.echo('---')
+                            click.echo("---")
                         click.echo(f"Time: {duration:.2f} seconds")
                         text = self.prompt_app.prompt(default=sql)
                     except KeyboardInterrupt:
@@ -927,7 +927,7 @@ def cli(
 
     litecli.connect(database)
 
-    litecli.logger.debug("Launch Params: \n" "\tdatabase: %r", database)
+    litecli.logger.debug("Launch Params: \n\tdatabase: %r", database)
 
     #  --execute argument
     if execute: