+++ tests/test_parseutils.py
@@ -120,12 +120,12 @@ def test_query_starts_with_comment():
 
 
 def test_queries_start_with():
-    sql = "# comment\n" "show databases;" "use foo;"
+    sql = "# comment\nshow databases;use foo;"
     assert queries_start_with(sql, ("show", "select")) is True
     assert queries_start_with(sql, ("use", "drop")) is True
     assert queries_start_with(sql, ("delete", "update")) is False
 
 
 def test_is_destructive():
-    sql = "use test;\n" "show databases;\n" "drop database foo;"
+    sql = "use test;\nshow databases;\ndrop database foo;"
     assert is_destructive(sql) is True